change stacks and dumps URLs
